Updates to searching to weight results based on the type of match
Search weighting is very nearly at the point where a database is
necessary. Micky from NZSL has been notified of this, but we can scrape
by with string matching for now.

The search previously did a contains match on main, secondary and Maori
gloss. This results in many inaccurate matches. The search now follows
the below rules:

* Exact matches on main or maori gloss are preferred and shown as first
  preference
* Inexact matches on main or maori gloss are shown as second preference
* Exact matches on secondary gloss are shown as third preference
* Inexact matches on secondary gloss are shown as fourth preference

This change improves search results, but still yields some strange
results with words matching when we would not expect, or vice-versa. I
expect further changes will be made based on exact search scenarios.