Stop using TreeMap
TreeMap is not the data structure we want, as it sorts by key, and keys must be unique.
We are wanting to order by weighting, so using this as they key, which results in a maximum
set capacity of _n_, where _n_ is the number of unique weightings we have.

Instead, we must take a more simple approach.