When entering the main activity, always hide the keyboard
This is always what we want, and it has the bonus of showing the whole
app when it first opens.

Without this, the keyboard starts visible, taking up up to half of the
app screen. This is pretty bad when the app opens, AND when returning
from a search result (since the search has already been performed)