Request new permission: Access Network State
This permission is required to detect whether the device the app is
running on has a network connection or not. If no network connection is
present, we do not try and load the video, and instead display a message
to the user.

Without this functionality, the videoview times out after a long time -
30 seconds to a minute, during which the videoview displays an empty
black screen.