Use LinkedHashSet for collecting results
Apparently Java is all about picking the right kind of collection. I'm
definitely not being very considerate of system resources here, but it
does work.

We were using an ArrayList to collate results, which does not enforce
any uniqueness. For search results that had signs that matched in more
than one way (e.g. equals, starts with), the sign would show up twice.
To fix this, I've moved to collating results to a LinkedHashSet, which
enforces uniqueness (first in wins). I experimented with passing the set
back to the activity, but this did not work as the Set interface does
not expose a method to get an item by index, and both the activity and
underlying tests use this method.