Use a stock MediaController instead of a 'no hide' option
Keeping the video controls visible at all times can be useful, but on
smaller screens, it overlays the video, obscuring the ability of a user
to actually see the sign video.

This change reverts to a normal media controller, which toggles the
display of the video controls on tap, meaning that while the video
controls continue to overlay the video, they can be "dismissed" to
display the full video.