Start playing the video as soon as it is ready
This should help to fix a UX issue where the video controls
are hidden by default, so unless a user knows to tap on the video
then tap the play icon, the video appears to not play.

To avoid this, we can just start playing the video once it's ready.